lantiq: switch to the mainline PCIe PHY driver
A driver for the PCIe PHY on the xRX200 SoCs is now sent upstream for
mainline inclusion. Switch the custom PCIe controller driver to use this
new PCIe PHY driver instead of the custom, hardcoded logic.

While here, also switch PCIe controller driver to parse devicetree for
the reset line and the RCU regmap. This ensures that the reset framework
can do the ref-counting for the shared reset line and allows the RCU
regmap to see all changes (instead of having some drivers updating the
registers by bypassing the regmap).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>