lantiq: dts: route the PCI interrupt through EBU
The PCI_INTA interrupt is not provided by ICU directly. Instead it is
routed from ICU to EBU which has registers to enable and configure (the
type of) the PCI_INTA interrupt.

Mark ebu0 as interrupt-controller and describe the actual PCI_INTA
interrupt routing in all .dts files.

Mark the EBU as interrupt-controller in Amazon-SE as well, even though
it does not have a PCI controller.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>