lantiq: dts: use the upstream dt-bindings for the ath5k EEPROM
Switch to the upstream device-tree bindings for ath5k. This means we can
now pass the EEPROM data as nvmem-cell and the MAC address either also
as nvmem-cell or (by using the OpenWrt custom device-tree parsing)
mtd-mac-address.

The "ath,eep-swap" is not needed anymore as ath5k can now auto-detect
when to apply the EEPROM byte-swapping.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>